+++ tendra/src/tools/tcc/list.c
@@ -95,7 +95,7 @@ new_list(void)
 
 	if (no_free == 0) {
 		no_free = 1000;
-		free_objs = alloc_nof(list, no_free);
+		free_objs = alloc_nof(list, (size_t) no_free);
 	}
 
 	return (free_objs + (--no_free));
